Beginning to refactor WiFiRanging code
