+++ app/src/main/java/uk/ac/warwick/cim/unheardCity/WifiScan.java
@@ -1,16 +1,26 @@
 package uk.ac.warwick.cim.unheardCity;
 
+import android.annotation.SuppressLint;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
+import android.content.pm.PackageManager;
 import android.net.wifi.ScanResult;
 import android.net.wifi.WifiManager;
+import android.net.wifi.rtt.RangingRequest;
+import android.net.wifi.rtt.RangingResult;
+import android.net.wifi.rtt.RangingResultCallback;
+import android.net.wifi.rtt.WifiRttManager;
+import android.os.Build;
 import android.os.Handler;
 import android.util.Log;
 
+import androidx.annotation.RequiresApi;
+
 import java.io.File;
 import java.util.List;
+import java.util.concurrent.Executor;
 
 
 public class WifiScan  implements Scan {
@@ -23,15 +33,18 @@ public class WifiScan  implements Scan {
 
     private Runnable wifiScanRunner;
 
+    private BroadcastReceiver wifiScanReceiver;
+
+    private BroadcastReceiver wifiRangeReceiver;
+
+    private WifiRttManager wifiRttManager;
+
     private Handler handler = new Handler();
 
     //OS throttles to 4 scans in 2 minutes so run at 30 seconds.
     private final static int timeInterval = 30000;
 
     private final static String TAG = "WiFiCScan";
-
-    private BroadcastReceiver wifiScanReceiver;
-
 
     protected WifiScan (Context context, WifiManager wManager, File file) {
         wifiManager = wManager;
@@ -102,6 +115,57 @@ public class WifiScan  implements Scan {
         // handle failure: new scan did NOT succeed
         // Provide a message
         Log.i("WIFI", "Scan wifi failed");
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.P)
+    private void startWiFiRanging() {
+        Log.i(TAG, "WiFi Ranging ON");
+
+        wifiRttManager = (WifiRttManager) ctx.getSystemService(Context.WIFI_RTT_RANGING_SERVICE);
+        if (ctx.getPackageManager().hasSystemFeature(PackageManager.FEATURE_WIFI_RTT)) {
+            //RTT available but is it enabled? User may have changed its state.
+            IntentFilter filter = new IntentFilter(wifiRttManager.ACTION_WIFI_RTT_STATE_CHANGED);
+
+            wifiRangeReceiver = new BroadcastReceiver() {
+                //Suppress the Fine Location permission because we request earlier to run the app.
+                @SuppressLint("MissingPermission")
+                @Override
+                public void onReceive(Context context, Intent intent) {
+                    if (wifiRttManager.isAvailable()) {
+                        List<ScanResult> results = wifiManager.getScanResults();
+                        RangingRequest.Builder builder = new RangingRequest.Builder();
+                        builder.addAccessPoints(results);
+                        RangingRequest req = builder.build();
+                        Executor executor = ctx.getMainExecutor();
+
+                        wifiRttManager.startRanging(req, executor, new RangingResultCallback() {
+
+                            @Override
+                            public void onRangingFailure(int code) {
+                                Log.i("WiFiRTT", "Ranging failure " + code);
+                            }
+
+                            @Override
+                            public void onRangingResults(List<RangingResult> results) {
+                                for (RangingResult result: results) {
+                                    //We have data, let's get some details
+                                    result.getDistanceMm();
+                                    result.getStatus();
+                                }
+                            }
+                        });
+                    } else {
+                        Log.i("WiFiRTT", "Ranging not available");
+                    }
+                }
+            };
+            ctx.registerReceiver(wifiRangeReceiver, filter);
+        }
+    }
+
+    private void stopWiFiRanging () {
+        Log.i(TAG, "WiFi Ranging OFF");
+        ctx.unregisterReceiver(wifiRangeReceiver);
     }
 
 }
